add samples
